+++ src/components/Section/index.ts
@@ -1,0 +1,2 @@
+export { Section } from "./Section";
+export type { SectionType, MainSection, ContentSection } from "./types";
